separate by geom position sorting
